requirements.txt
flask==3.0.0
aiohttp==3.9.1
pandas==2.1.4
openpyxl==3.1.2
beautifulsoup4==4.12.2
lxml==5.1.0
requests==2.31.0
python-dotenv==1.0.0
playwright
playwright-stealth
